requirements.txt
Django~=3.2.2
django-jsonfield~=1.4.1
django-multiselectfield~=0.1.12
djangorestframework~=3.12.2
django-rest-swagger~=2.2.0
drf-nested-routers~=0.93.3
drf-yasg~=1.20.0
ensembl-prodinf-djcore>=1.2.0<1.3.0
django-ckeditor~=6.0.0
